+++ R/CohortDiagnostics.R
@@ -1,4 +1,4 @@
-# Copyright 2022 Observational Health Data Sciences and Informatics
+# Copyright 2023 Observational Health Data Sciences and Informatics
 #
 # This file is part of CohortDiagnostics
 #
@@ -21,7 +21,7 @@
 #' @importFrom grDevices rgb
 #' @importFrom stats aggregate
 #' @importFrom utils write.csv install.packages menu unzip setTxtProgressBar txtProgressBar packageName
-#' @import dplyr
+#' @importFrom dplyr %>%
 #' @importFrom rlang .data
 #' @importFrom methods is
 #' @importFrom FeatureExtraction createDefaultCovariateSettings createTemporalCovariateSettings